Work in fullscreen and full resolution

+++ src/input.h
@@ -33,8 +33,8 @@ class Mouse {
   static bool is_btn_down(int btn);
   static bool is_btn_pressed(int btn);
   static bool is_btn_released(int btn);
-  static int get_x();
-  static int get_y();
+  static double get_x();
+  static double get_y();
   /*bool is_button_down(int key);
   bool is_button_pressed(int key);
   bool is_button_released(int key);*/
@@ -44,7 +44,7 @@ class Mouse {
 
   static void cursor_cb(GLFWwindow* wnd, double x, double y);
   static void mouse_cb(GLFWwindow* wnd, int button, int action, int mods);
-  static int m_x, m_y;
+  static double m_x, m_y;
   static std::map<int, bool> press_btns;
   static std::map<int, bool> release_btns;
   static std::map<int, bool> repeat_btns;
